+++ libs/application/templates/login-service/src/fields/IsatInvalid/IsatInvalid.tsx
@@ -13,36 +13,34 @@ import { errorMessages } from '../../lib/messages'
 export const IsatInvalid = ({ application }: FieldBaseProps) => {
   const { formatMessage } = useLocale()
   return (
-    <>
-      <GridColumn span={['1/1', '1/1', '1/1']} paddingTop={2}>
-        <AlertMessage
-          type="error"
-          title={formatText(
-            errorMessages.invalidIsatMessage,
-            application,
-            formatMessage,
-          )}
-          message={
-            <Box component="span" display="block">
-              <Link
-                href={formatText(
+    <GridColumn span="1/1" paddingTop={2}>
+      <AlertMessage
+        type="error"
+        title={formatText(
+          errorMessages.invalidIsatMessage,
+          application,
+          formatMessage,
+        )}
+        message={
+          <Box component="span" display="block">
+            <Link
+              href={formatText(
+                errorMessages.invalidIsatPdfUrl,
+                application,
+                formatMessage,
+              )}
+            >
+              <Text variant="small">
+                {formatText(
                   errorMessages.invalidIsatPdfUrl,
                   application,
                   formatMessage,
                 )}
-              >
-                <Text variant="small">
-                  {formatText(
-                    errorMessages.invalidIsatPdfUrl,
-                    application,
-                    formatMessage,
-                  )}
-                </Text>
-              </Link>
-            </Box>
-          }
-        />
-      </GridColumn>
-    </>
+              </Text>
+            </Link>
+          </Box>
+        }
+      />
+    </GridColumn>
   )
 }
