--- libs/application/templates/login-service/src/shared/types.ts
@@ -1,7 +1,0 @@
-export type selectCompanySearchField = {
-  searchField: {
-    isat: string
-    nationalId: string
-    label: string
-  }
-}
