+++ libs/application/ui-components/src/index.ts
@@ -11,3 +11,4 @@ export { formatBankInfo, formatPhoneNumber } from './utilities/formatters'
 export { handleServerError } from './utilities/handleServerError'
 export { InputImageUpload } from './components/InputImageUpload/InputImageUpload'
 export { PaymentPending } from './components/PaymentPending/PaymentPending'
+export * from './types'
