+++ libs/application/ui-components/src/types.ts
@@ -1,0 +1,7 @@
+export type selectCompanySearchField = {
+  searchField: {
+    isat: string
+    nationalId: string
+    label: string
+  }
+}
